requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==109.0.5414.25.0
click==8.1.3
colorama==0.4.5
confection==0.0.3
cymem==2.0.7
exceptiongroup==1.0.4
filelock==3.8.0
Flask==2.2.2
Flask-Markdown==0.3
frozendict==2.3.4
gdown==4.5.1
greenlet==2.0.1
h11==0.14.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
murmurhash==1.0.9
NERP==1.0.2.2
nlpaug==1.1.11
nltk==3.7
numpy==1.23.2
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pathy==0.10.1
pluggy==1.0.0
preshed==3.0.8
progressbar==2.5
pyconll==3.1.0
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
selenium==4.7.2
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
smart-open==6.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy-legacy==3.0.10
spacy-loggers==1.0.4
SQLAlchemy==1.4.44
srsly==2.4.5
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
TorchCRF==1.1.0
tqdm==4.64.0
transformers==4.21.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.3.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
